ADD: user-registration

+++ studybud/base/views.py
@@ -1,6 +1,9 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.http import HttpResponse
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from django.db.models import Q
 from .models import Room, Topic
 from django.contrib.auth.models import User
@@ -32,6 +35,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -43,9 +47,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form =RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Permission Denied')
+
     if request.method == 'POST':
         # for update, we need to pass the instance of the room
         form = RoomForm(request.POST,instance=room)
@@ -56,8 +65,13 @@ def updateRoom(request,pk):
     context ={'form':form}
     return render(request,'room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Permission Denied')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
@@ -66,8 +80,13 @@ def deleteRoom(request,pk):
     return render(request,'delete.html',context)
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
+        username = username.lower()
         password = request.POST.get('password')
 
         try:
@@ -85,9 +104,25 @@ def loginPage(request):
         else:
             messages.error(request,'Username OR password is incorrect')
 
-    context = {}
+    context = {'page':page}
     return render(request,'login_register.html',context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    context = {'form':form}
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False) # we don't want to save the user yet
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured during registration')  
+    return render(request,'login_register.html',context)
